test: Add tests for non-root execution of monitor
Add e2e tests for nerdctl, crictl and docker for testing the non-root
execution of the unikernel. In partiuclar, in the cases we need KVM
access (i.e. Hvt, Qemu, FC) we get the group ID of KVM and then add it
as additional group and we try to execute the unikernel as nobody.

Regarding the test check, we get the PID of the monitor process and
check the values in the PID, GID and Groups fields. These values should
match the arguments we gave in the test description.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>