feat(urunc-deploy): Implement urunc-deploy
to facilicate the deployment of urunc and its dependencies on a k8s cluster

This component consists of the necessary k8s manifests to deploy urunc and a
container image containing the necessary hypervisor binaries and the Bash script
that installs them on the host andconfigures containerd to support urunc. When
the urunc-deploy component is deployed, it willspawn a privileged Pod that runs
the install.sh script to install the necessary hypervisor binaries
under /usr/local/bin and configure containerd to support urunc.
If QEMU is installed by urunc-deploy,the QEMU BIOSes will be copied in
the host machine and urunc-deploy will create
a Bash wrapper for QEMU to be able to locate the BIOS files in the host machine.
The urunc-deploy component also includes a cleanup script that removes the
hypervisor binaries and the containerd configuration.

PR: https://github.com/nubificus/urunc/pull/91
Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>