fix(knative): Fix readiness probe var for Knative
When we inject the REDIRECT_IP variable in the queue-proxy
podspec to point to the unikernel endpoint, we assumed that
the port of the user-container would be always 8080. This patch
fixes that, allowing the user to define the port that the unikernel
container will listen to.

PR: https://github.com/nubificus/urunc/pull/146
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>