ci(deploy): Fix generic Incus client name conflict
When spawning multiple concurrent runners that use the
incus-remote-setup-action, the client name was shared
across the runners. As a result, the first runner to reach the
cleanup step would mark the client as unauthorized for all other
runners. This commit fixes the issue by using a unique
name for each client.

PR: https://github.com/nubificus/urunc/pull/147
Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>