refactor(cmdline): Change the way we handle the unikernel cmdline
Instead of combining all cmd arguments in one string and passing it to
the unikernel, keep the cmd as an array of strings and let each
implementation of the unikernel interface to handle it. This will be
required for Linux support and handling of multi-word arguments.

PR: https://github.com/nubificus/urunc/pull/158
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>