feat(chroot): Add support for pivot/chroot before monitor execution
Instead of using the host's rootfs, create a new one based on the
container's rootfs and invoke the monitor execution from inside the new
rootfs (after a chroot/pivot).

To make this possible, we perform the necessary preparation steps for
the container's rootfs (propagation etc.) and apend the necessary
files/device for the monitor execution. In particular, we add support
for bind mounting or copying the monitor binary and also the creation of
the KVM, null, urandom, TUNTAP devices.

Furthermore, such design allows us to play around with permissions and
hence we set rw in KVM/TUNTAP and devmapper's block in order to easily
let the monitor process access these devices without any requirement of
specifying groupids etc. This is safe, sicne we have a single process
runing inside the container and it is the monitor that already requires
access to these devices.

PR: https://github.com/urunc-dev/urunc/pull/187
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>