feat(chroot): Mount shared objects from host for monitor execution
Qemu and Solo5 are typically built as dynamically linked binaries and
hence they require some shared objects. Unfortunately, it is too
cumbersome to manually identify the dependencies and mount them inside
the new rootfs. Instead, we opt for mounting all /lib, lib63 and
/usr/lib, since these are the directories that usually host such share
objects.

This is not good, but useful as a first step in chroot and to avoid an
abrupt transition to static monitor binaries. We should remove this
commit as soon as we add support for statically linked binaries.

PR: https://github.com/urunc-dev/urunc/pull/187
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>