fix(unikontainers): Refactor sequential hook execution
Refactor the sequential execution of hooks in the unikontainers package
to allow logging of each hook's potential errors. This change
enhances the visibility of issues that may arise during the execution
of hooks, making it easier to debug and maintain the code.
This commit also improves the handling of the waitgroup used for
hook execution and initializes goroutines in a cleaner way.

PR: https://github.com/urunc-dev/urunc/pull/190
Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Reviewed-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>