fix(urunc-deploy): Change qemu handling in urunc-deploy
Since, we do not have support for bash inside the new monitor rootfs
anymore, we can not use the hack with the bash script. Instead, directly
set the location of the bios/data files of qemu from urunc. Also, update
urunc-deploy to use that exact directory for installing bios/data files
and not to create the qemu script.

PR: https://github.com/urunc-dev/urunc/pull/187
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>