fix(logs): Fix log levels
This commit fixes the log levels of urunc to ensure that the correct
severity levels are used for logging messages. This change improves the
clarity and consistency of log output, making it easier to identify
issues and understand the state of the system during operation.

PR: https://github.com/urunc-dev/urunc/pull/190
Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Reviewed-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>